feat: Убрать бизнес-логику из мапперов
