+++ src/main/java/ru/practicum/shareit/booking/mapper/BookingDtoMapper.java
@@ -1,36 +1,27 @@
-package ru.practicum.shareit.booking.dto;
+package ru.practicum.shareit.booking.mapper;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import ru.practicum.shareit.booking.dto.BookingDto;
+import ru.practicum.shareit.booking.dto.BookingResultDto;
 import ru.practicum.shareit.booking.model.Booking;
 import ru.practicum.shareit.booking.model.BookingStatus;
 import ru.practicum.shareit.item.mapper.ItemDtoMapper;
 import ru.practicum.shareit.item.model.Item;
-import ru.practicum.shareit.item.repository.ItemRepository;
-import ru.practicum.shareit.user.dto.UserDtoMapper;
+import ru.practicum.shareit.user.mapper.UserDtoMapper;
 import ru.practicum.shareit.user.model.User;
-import ru.practicum.shareit.user.repository.UserRepository;
 
 @Slf4j
 @Component
 public class BookingDtoMapper {
 
     private static final String OBJECT_MAPPED_MESSAGE = "Выполнено преобразование объекта из {} в {}";
-    private final ItemRepository itemRepository;
-    private final UserRepository userRepository;
     private final UserDtoMapper userDtoMapper;
     private final ItemDtoMapper itemDtoMapper;
 
     @Autowired
-    public BookingDtoMapper(
-            ItemRepository itemRepository,
-            UserRepository userRepository,
-            UserDtoMapper userDtoMapper,
-            ItemDtoMapper itemDtoMapper
-    ) {
-        this.itemRepository = itemRepository;
-        this.userRepository = userRepository;
+    public BookingDtoMapper(UserDtoMapper userDtoMapper, ItemDtoMapper itemDtoMapper) {
         this.userDtoMapper = userDtoMapper;
         this.itemDtoMapper = itemDtoMapper;
     }
@@ -41,20 +32,8 @@ public class BookingDtoMapper {
                 booking.getStart(),
                 booking.getEnd(),
                 booking.getStatus(),
-                userDtoMapper.mapUserToDto(userRepository.getReferenceById(booking.getBooker().getId())),
-                itemDtoMapper.mapItemToDto(itemRepository.getReferenceById(booking.getItem().getId()))
-        );
-        log.trace(OBJECT_MAPPED_MESSAGE, booking, resultDto);
-        return resultDto;
-    }
-
-    public BookingDto mapBookingToBookingDto(Booking booking) {
-        BookingDto resultDto = new BookingDto(
-                booking.getId(),
-                booking.getItem().getId(),
-                booking.getStart(),
-                booking.getEnd(),
-                booking.getBooker().getId()
+                userDtoMapper.mapUserToDto(booking.getBooker()),
+                itemDtoMapper.mapItemToDto(booking.getItem())
         );
         log.trace(OBJECT_MAPPED_MESSAGE, booking, resultDto);
         return resultDto;
